% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate_boards.R
\name{activate_boards}
\alias{activate_boards}
\title{Activate main and current boards for a given study_run}
\usage{
activate_boards(study_run)
}
\arguments{
\item{study_run}{A character string indicating the specific study run to activate.}
}
\value{
A list with two elements: \code{main} (main board object) and \code{current} (requested board object).
}
\description{
This function reads the board index and validates the structure of the study.
It ensures that a given `study_id` has exactly one main, one first, and at least one rerun.
It also ensures that the requested `study_run` is valid.
}
\examples{
boards <- activate_boards("01A_rerun2")
pins::pin_write(board = boards$main, x = df, name = "my_data", type = "qs")
}
